fix(ci): isort black profile, Java test inner classes, Jest tsconfig
CI: isort --profile black for black-compatible import sorting.

Java tests: CircuitBreakerConfig -> CircuitBreaker.CircuitBreakerConfig,
BulkheadConfig -> Bulkhead.BulkheadConfig, ambiguous assertEquals cast,
missing java.util.List import in TypesTest.

JS: Jest ts-jest transform overrides tsconfig rootDir/exclude so
test files in tests/ can import from src/.

+++ sdks/java/aether-sdk/src/test/java/io/aether/sdk/state/StateHandleTest.java
@@ -171,6 +171,6 @@ class StateHandleTest {
 
         StateHandle s2 = new StateHandle("a", CapabilitySet.all(), backend);
         s2.load().join();
-        assertEquals(42, s2.get("x"));
+        assertEquals(42, (int) s2.get("x"));
     }
 }
